# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel API",
        "description": "API documentation for the Hotel Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/room-types": {
            "get": {
                "tags": [
                    "RoomTypes"
                ],
                "summary": "Get all room types (paginated, filterable by id)",
                "operationId": "ca013b8668b9235311faca0f7063a058",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Filter by room type id (single or array)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of room types"
                    }
                }
            },
            "post": {
                "tags": [
                    "RoomTypes"
                ],
                "summary": "Create a new room type",
                "operationId": "45965ec37cb311105c9806350f4b9e21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "external_id",
                                    "name"
                                ],
                                "properties": {
                                    "external_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room type created"
                    }
                }
            }
        },
        "/api/room-types/{id}": {
            "get": {
                "tags": [
                    "RoomTypes"
                ],
                "summary": "Get a room type by ID",
                "operationId": "32dfe7550e8eb4648f15f4037cca60ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room type details"
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "RoomTypes"
                ],
                "summary": "Update a room type",
                "operationId": "cb1bc264a4c17bd1f46e409a03526d6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room type updated"
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "RoomTypes"
                ],
                "summary": "Delete a room type",
                "operationId": "53dbaa498cdf5e36a364a6cb19d725f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Room type deleted"
                    },
                    "404": {
                        "description": "Room type not found"
                    }
                }
            }
        }
    },
...
}